moving files and create RemoveDatabaseSettings
