+++ Settings/ConnectionStrings/Database/RemoveDatabaseSettings.cs
@@ -1,0 +1,6 @@
+﻿using Spectre.Console.Cli;
+
+namespace config.Settings.ConnectionStrings.Database;
+internal class RemoveDatabaseSettings : CommandSettings
+{
+}
